ref: improved time-to-first-element by moving Chroma logic around
